refactor(main): use a unexported status struct, switch to json/v2
In the main package monitor.go uses its own version of the wifire.Status struct
so we can track time as floats to simplify the prediction code.

Switching the json/v2 adds time.Duration support, so kill off the JSONDuration
type. This also triggers a bug in go-vulncheck so it panics, see:
https://github.com/golang/go/issues/73871